community component refactor and button functionality
