+++ client/src/pages/Communities.jsx
@@ -18,6 +18,7 @@ export default function Communities() {
   const [joinCommunity, { error }] = useMutation(JOIN_COMMUNITY, {
     refetchQueries: [
       { query: DASHBOARD_QUERY, variables: { username: user, sort } },
+      { query: QUERY_COMMUNITIES },
     ],
   });
 
@@ -41,10 +42,6 @@ export default function Communities() {
 
   return (
     <div className="flex flex-col items-center">
-      <h2 className="mt-4 text-5xl font-bold text-zinc-200">
-        {/* Welcome {user.username}! */}
-      </h2>
-
       <div className="mt-4 flex flex-row justify-evenly">
         <CommunityList communities={communities} sort={setSort} join={onJoin} />
       </div>
